Major Update: All previous nodes' programs must be upgraded to the latest version to use. This update adds SafeAsset, SafeApp, Candy Protocol, and Token Locking Feature.

+++ src/masternodeman.cpp
@@ -711,7 +711,7 @@ bool CMasternodeMan::GetMasternodeByRank(int nRankIn, masternode_info_t& mnInfoR
     if (!GetMasternodeScores(nBlockHash, vecMasternodeScores, nMinProtocol))
         return false;
 
-    if (vecMasternodeScores.size() < nRankIn)
+    if ((int)vecMasternodeScores.size() < nRankIn)
         return false;
 
     int nRank = 0;
